requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
distro==1.9.0
easyocr==1.7.2
fastapi==0.123.5
filelock==3.20.0
fsspec==2025.12.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
jiter==0.12.0
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6
ninja==1.13.0
numpy==2.2.6
openai==2.8.1
opencv-python-headless==4.12.0.88
packaging==25.0
pdf2image==1.17.0
pillow==12.0.0
pyclipper==1.4.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.26.6
python-bidi==0.6.7
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
scikit-image==0.25.2
scipy==1.16.3
shapely==2.1.2
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
tifffile==2025.10.16
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uv==0.9.15
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
